+++ .rubocop.yml
@@ -48,6 +48,12 @@ Style/AutoResourceCleanup:
 Style/CollectionMethods:
   Enabled: true
 
+Style/Documentation:
+  Enabled: false
+  Exclude:
+    - 'lib/pwnlib/shellcraft/generators/**/*.rb'
+    - 'test/*.rb'
+
 Style/Encoding:
   Enabled: false
