+++ lib/pwnlib/shellcraft/generators/amd64/infloop.rb
