+++ lib/pwnlib/shellcraft/generators/amd64/linux/ls.rb
