+++ lib/pwnlib/shellcraft/generators/amd64/linux/sh.rb
