+++ lib/pwnlib/shellcraft/generators/amd64/linux/syscalls/syscall.rb
