+++ lib/pwnlib/shellcraft/generators/amd64/nop.rb
