+++ lib/pwnlib/shellcraft/generators/amd64/pushstr.rb
