+++ lib/pwnlib/shellcraft/generators/amd64/pushstr_array.rb
