+++ lib/pwnlib/shellcraft/generators/amd64/setregs.rb
