+++ lib/pwnlib/shellcraft/generators/i386/infloop.rb
