+++ lib/pwnlib/shellcraft/generators/i386/linux/ls.rb
