+++ lib/pwnlib/shellcraft/generators/i386/linux/sh.rb
