+++ lib/pwnlib/shellcraft/generators/i386/linux/syscalls/syscall.rb
