+++ lib/pwnlib/shellcraft/generators/i386/nop.rb
