+++ lib/pwnlib/shellcraft/generators/i386/pushstr.rb
