+++ lib/pwnlib/shellcraft/generators/i386/pushstr_array.rb
