+++ lib/pwnlib/shellcraft/generators/i386/setregs.rb
