+++ lib/pwnlib/shellcraft/generators/x86/common/common.rb
@@ -1,0 +1,32 @@
+require 'pwnlib/context'
+require 'pwnlib/shellcraft/generators/amd64/common/mov'
+require 'pwnlib/shellcraft/generators/amd64/common/pushstr'
+require 'pwnlib/shellcraft/generators/helper'
+require 'pwnlib/shellcraft/generators/i386/common/mov'
+require 'pwnlib/shellcraft/generators/i386/common/pushstr'
+
+module Pwnlib
+  module Shellcraft
+    module Generators
+      module X86
+        # For non os-related methods.
+        module Common
+          module_function
+
+          %i[mov pushstr].each do |m|
+            define_method(m) do |*args|
+              if context.arch == 'amd64'
+                Generators::Amd64::Common.send(m, *args)
+              elsif context.arch == 'i386'
+                Generators::I386::Common.send(m, *args)
+              end
+            end
+          end
+
+          extend Helper
+          include ::Pwnlib::Context
+        end
+      end
+    end
+  end
+end
