+++ lib/pwnlib/shellcraft/generators/x86/linux/ls.rb
