+++ lib/pwnlib/shellcraft/generators/x86/linux/sh.rb
