+++ lib/pwnlib/shellcraft/generators/x86/linux/syscall.rb
