--- lib/pwnlib/shellcraft/templates/x86/infloop.rb
@@ -1,5 +1,0 @@
-::Pwnlib::Shellcraft.define(__FILE__) do
-  label = get_label('infloop')
-  cat "#{label}:"
-  cat "jmp #{label}"
-end
